test(wow-compensation-domain): move ExecutionFailedTest to ExecutionFailedSpec
- Transfer all test cases from ExecutionFailedTest to ExecutionFailedSpec
- Remove ExecutionFailedTest file
- Update related imports in ExecutionFailedSpec
